Fix devname overflow
When the dev name is longer than the buffer, no NUL byte is appended
which can cause issues like printing garbage in the Xorg log file. Since
names of HID devices (struct hid_device) can be of length 128, bump the
length accordingly and make place for a NUL byte (this NUL byte is set
with the memset a few lines above the ioctl).

Signed-off-by: Peter Wu <lekensteyn@gmail.com>