First usable version of coasting and smooth scrolling features.
It lacks configuration, needs cleaning, and may not perfect as rest of my code ;) but hey is't here.

Scroll gesture calculates speed of scroll, which is then translated to scroll xy movement during regular
scroll gesture and after that during coasting.
Default friction factor for coasting is 0.7, which is not configurable right now, so it's not final commit.
Update delta time is set to 20 miliseconds.

Rename 'high precision scrolling' to 'smooth scroll'.
Remove scroll 'stepping' in smooth mode.
Coasing is availible for any kind of swipe that normally would generate button {4-7} events.